src/main/java/com/hcl/calculator/application/SimpleCalculatorDTO.java
package com.hcl.calculator.application;

public class SimpleCalculatorDTO {
    private double firstValue;
    private double secondValue;
    private String operator;

    public double getFirstValue() {
        return firstValue;
    }

    public void setFirstValue(double firstValue) {
        this.firstValue = firstValue;
    }

    public double getSecondValue() {
        return secondValue;
    }

    public void setSecondValue(double secondValue) {
        this.secondValue = secondValue;
    }

    public String getOperator() {
        return this.operator;
    }

    public void setOperator(final String operator) {
        this.operator = operator;
    }
}
